requirements.txt
fastapi
uvicorn
sentence-transformers
python-dotenv
faiss-cpu
passlib[bcrypt]==1.7.4
pyjwt
python-multipart
llama-cpp-python==0.2.83
huggingface-hub
langchain-core
supabase
pypdf
requests
bcrypt==4.0.1